// frontend/src/app/api/generated/model/customerSearchKind.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export enum CustomerSearchKind {
  CUSTOMER_NUMBER = 'CUSTOMER_NUMBER',

  EMAIL = 'EMAIL',

  PHONE = 'PHONE',

  NAME = 'NAME',
}